+++ fastapi_mvc/generators/generator/template/{{cookiecutter.folder_name}}/lib/generators/{{cookiecutter.generator_name}}/__init__.py
@@ -2,5 +2,5 @@
 from .{{cookiecutter.generator_name}} import {{cookiecutter.class_name}}
 
 # NOTE! Do not edit this! Method for programmatically loading user generators
-# depends on having only one class in module __all__ attribute.
-__all__ = {{cookiecutter.class_name}}
+# depends on having only one class in module `generator_class` attribute.
+generator_class = {{cookiecutter.class_name}}