+++ tests/unit/data/lib/generators/foobar/__init__.py
@@ -2,5 +2,5 @@
 from .foobar import FoobarGenerator
 
 # NOTE! Do not edit this! Method for programmatically loading user generators
-# depends on having only one class in module __all__ attribute.
-__all__ = FoobarGenerator
+# depends on having only one class in module `generator_class` attribute.
+generator_class = FoobarGenerator