+++ tests/unit/data/lib/generators/my_controller/__init__.py
@@ -2,5 +2,5 @@
 from .my_controller import MyControllerGenerator
 
 # NOTE! Do not edit this! Method for programmatically loading user generators
-# depends on having only one class in module __all__ attribute.
-__all__ = MyControllerGenerator
+# depends on having only one class in module `generator_class` attribute.
+generator_class = MyControllerGenerator